server/pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["tc-imba <liuyh970615@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.70.1"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
pydantic-universal-settings = "^0.1.5"
sqlmodel = "^0.0.5"
aiosqlite = "^0.17.0"
starlette-context = "^0.3.3"
SQLAlchemy-Utils = "^0.37.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
